+++ app/models/table.rb
@@ -5,6 +5,7 @@ require 'forwardable'
 require_relative './table/column_typecaster'
 require_relative './table/privacy_manager'
 require_relative './table/relator'
+require_relative './table/user_table'
 require_relative './visualization/member'
 require_relative './visualization/collection'
 require_relative './visualization/overlays'
@@ -16,30 +17,16 @@ require_relative '../../services/importer/lib/importer/query_batcher'
 require_relative '../../services/datasources/lib/datasources/decorators/factory'
 require_relative '../../services/table-geocoder/lib/internal-geocoder/latitude_longitude'
 
-class Table < Sequel::Model(:user_tables)
+class Table
   extend Forwardable
-
-  PRIVACY_PRIVATE = 0
-  PRIVACY_PUBLIC = 1
-  PRIVACY_LINK = 2
-
-  PRIVACY_PRIVATE_TEXT = 'private'
-  PRIVACY_PUBLIC_TEXT = 'public'
-  PRIVACY_LINK_TEXT = 'link'
-
-  PRIVACY_VALUES_TO_TEXTS = {
-      PRIVACY_PRIVATE => PRIVACY_PRIVATE_TEXT,
-      PRIVACY_PUBLIC => PRIVACY_PUBLIC_TEXT,
-      PRIVACY_LINK => PRIVACY_LINK_TEXT
-  }
 
   SYSTEM_TABLE_NAMES = %w( spatial_ref_sys geography_columns geometry_columns raster_columns raster_overviews cdb_tablemetadata geometry raster )
 
+   # TODO Part of a service along with schema
   CARTODB_COLUMNS = %W{ cartodb_id created_at updated_at the_geom }
   THE_GEOM_WEBMERCATOR = :the_geom_webmercator
   THE_GEOM = :the_geom
 
-  RESERVED_TABLE_NAMES = %W{ layergroup all }
 
   # @see services/importer/lib/importer/column.rb -> RESERVED_WORDS
   # @see config/initializers/carto_db.rb -> POSTGRESQL_RESERVED_WORDS & RESERVED_COLUMN_NAMES
@@ -65,21 +52,44 @@ class Table < Sequel::Model(:user_tables)
 
   DEFAULT_DERIVED_VISUALIZATION_POSTFIX = 'Map'
 
-  # Associations
-  many_to_one  :map
-  many_to_many :layers, join_table: :layers_user_tables,
-                        left_key:   :user_table_id,
-                        right_key:  :layer_id,
-                        reciprocal: :user_tables
-  one_to_one   :automatic_geocoding
-  one_to_many  :geocodings
-
-  plugin :association_dependencies, map:                  :destroy,
-                                    layers:               :nullify,
-                                    automatic_geocoding:  :destroy
-  plugin :dirty
 
   def_delegators :relator, *CartoDB::TableRelator::INTERFACE
+  def_delegators :@user_table, *::UserTable::INTERFACE
+
+
+  def initialize(args = {})
+    if args[:user_table].nil?
+      @user_table = UserTable.new(args)
+    else
+      @user_table = args[:user_table]
+    end
+    @user_table.set_service(self)
+  end
+
+  # forwardable does not work well with this one
+  def layers
+    @user_table.layers
+  end
+
+  def save
+    @user_table.save
+    self
+  end
+
+  def update(args)
+    @user_table.update(args)
+    self
+  end
+
+  def reload
+    @user_table.reload
+    self
+  end
+
+  # ----------------------------------------------------------------------------
+
+
+
 
   def public_values(options = {}, viewer_user=nil)
     selected_attrs = if options[:except].present?
@@ -96,9 +106,9 @@ class Table < Sequel::Model(:user_tables)
     attrs
   end
 
-  def default_privacy_values
-    self.owner.try(:private_tables_enabled) ? PRIVACY_PRIVATE : PRIVACY_PUBLIC
-  end #default_privacy_values
+  def default_privacy_value
+    self.owner.try(:private_tables_enabled) ? UserTable::PRIVACY_PRIVATE : UserTable::PRIVACY_PUBLIC
+  end
 
   def geometry_types_key
     @geometry_types_key ||= "#{key}:geometry_types"
@@ -131,31 +141,6 @@ class Table < Sequel::Model(:user_tables)
     schema.select { |key, value| value == 'raster' }.length > 0
   end
 
-  def_dataset_method(:search) do |query|
-    conditions = <<-EOS
-      to_tsvector('english', coalesce(name, '') || ' ' || coalesce(description, '')) @@ plainto_tsquery('english', ?) OR name ILIKE ?
-      EOS
-    where(conditions, query, "%#{query}%")
-  end
-
-  def_dataset_method(:multiple_order) do |criteria|
-    if criteria.nil? || criteria.empty?
-      order(:id)
-    else
-      order_params = criteria.map do |key, order|
-        Sequel.send(order.to_sym, key.to_sym)
-      end
-      order(*order_params)
-    end
-  end #multiple_order
-
-
-  # Ignore mass-asigment on not allowed columns
-  self.strict_param_setting = false
-
-  # Allowed columns
-  set_allowed_columns(:privacy, :tags, :description)
-
   attr_accessor :force_schema,
                 :import_from_file,
                 :import_from_url,
@@ -178,7 +163,7 @@ class Table < Sequel::Model(:user_tables)
     table = nil
     return table unless viewer_user
 
-    table_temp = Table.where(id: table_id).first
+    table_temp = UserTable.where(id: table_id).first.service
     unless table_temp.nil?
       vis = CartoDB::Visualization::Collection.new.fetch(
           user_id: viewer_user.id,
@@ -204,7 +189,7 @@ class Table < Sequel::Model(:user_tables)
           user_id = owner.id
         end
       end
-      Table.where(user_id: user_id, name: table_name).first
+      UserTable.where(user_id: user_id, name: table_name).first
     }
   end #tables_from
 
@@ -239,7 +224,7 @@ class Table < Sequel::Model(:user_tables)
     table = vis.nil? ? nil : vis.table
 
     if rx.match(id_or_name) && table.nil?
-      table_temp = Table.where(id: id_or_name).first
+      table_temp = UserTable.where(id: id_or_name).first.try(:service)
       unless table_temp.nil?
         # Make sure we're allowed to see the table
         vis = CartoDB::Visualization::Collection.new.fetch(
@@ -268,40 +253,24 @@ class Table < Sequel::Model(:user_tables)
 
   # Core validation method that is automatically called before create and save
   def validate
-    super
-
     ## SANITY CHECKS
-
-    # userid and table name tuple must be unique
-    validates_unique [:name, :user_id], :message => 'is already taken'
-
-    # tables must have a user
-    errors.add(:user_id, "can't be blank") if user_id.blank?
-
-    errors.add(
-      :name, 'is a reserved keyword, please choose a different one'
-    ) if RESERVED_TABLE_NAMES.include?(self.name) 
-
-    # privacy setting must be a sane value
-    if privacy != PRIVACY_PRIVATE && privacy != PRIVACY_PUBLIC && privacy != PRIVACY_LINK
-      errors.add(:privacy, "has an invalid value '#{privacy}'")
-    end
+    # TODO this should be fully moved to storage
 
     # Branch if owner does not have private table privileges
     unless self.owner.try(:private_tables_enabled)
       # If it's a new table and the user is trying to make it private
-      if self.new? && privacy == PRIVACY_PRIVATE
-        errors.add(:privacy, 'unauthorized to create private tables')
+      if self.new? && @user_table.privacy == UserTable::PRIVACY_PRIVATE
+        @user_table.errors.add(:privacy, 'unauthorized to create private tables')
       end
 
       # if the table exists, is private, but the owner no longer has private privileges
-      if !self.new? && privacy == PRIVACY_PRIVATE && self.changed_columns.include?(:privacy)
-        errors.add(:privacy, 'unauthorized to modify privacy status to private')
+      if !self.new? && @user_table.privacy == UserTable::PRIVACY_PRIVATE && @user_table.changed_columns.include?(:privacy)
+        @user_table.errors.add(:privacy, 'unauthorized to modify privacy status to private')
       end
 
       # cannot change any existing table to 'with link'
-      if !self.new? && privacy == PRIVACY_LINK && self.changed_columns.include?(:privacy)
-        errors.add(:privacy, 'unauthorized to modify privacy status to pubic with link')
+      if !self.new? && @user_table.privacy == UserTable::PRIVACY_LINK && @user_table.changed_columns.include?(:privacy)
+        @user_table.errors.add(:privacy, 'unauthorized to modify privacy status to pubic with link')
       end
 
     end
@@ -310,8 +279,7 @@ class Table < Sequel::Model(:user_tables)
   # runs before each validation phase on create and update
   def before_validation
     # ensure privacy variable is set to one of the constants. this is bad.
-    self.privacy ||= (owner.try(:private_tables_enabled) ? PRIVACY_PRIVATE : PRIVACY_PUBLIC)
-    super
+    @user_table.privacy ||= (owner.try(:private_tables_enabled) ? UserTable::PRIVACY_PRIVATE : UserTable::PRIVACY_PUBLIC)
   end
 
   def append_from_importer(new_table_name, new_schema_name)
@@ -414,7 +382,7 @@ class Table < Sequel::Model(:user_tables)
   end
 
   def import_to_cartodb(uniname=nil)
-    @data_import ||= DataImport.where(id: data_import_id).first || DataImport.new(user_id: owner.id)
+    @data_import ||= DataImport.where(id: @user_table.data_import_id).first || DataImport.new(user_id: owner.id)
     if migrate_existing_table.present? || uniname
       @data_import.data_type = DataImport::TYPE_EXTERNAL_TABLE
       @data_import.data_source = migrate_existing_table || uniname
@@ -511,17 +479,15 @@ class Table < Sequel::Model(:user_tables)
 
   def before_create
     raise CartoDB::QuotaExceeded if owner.over_table_quota?
-    super
-    update_updated_at
 
     # The Table model only migrates now, never imports
     if migrate_existing_table.present?
-      if self.data_import_id.nil? #needed for non ui-created tables
+      if @user_table.data_import_id.nil? #needed for non ui-created tables
         @data_import  = DataImport.new(:user_id => self.user_id)
         @data_import.updated_at = Time.now
         @data_import.save
       else
-        @data_import  = DataImport.find(:id=>self.data_import_id)
+        @data_import  = DataImport.find(:id=>@user_table.data_import_id)
       end
 
       importer_result_name = import_to_cartodb(name)
@@ -554,7 +520,6 @@ class Table < Sequel::Model(:user_tables)
   end
 
   def after_create
-    super
     self.create_default_map_and_layers
     self.create_default_visualization
 
@@ -565,8 +530,8 @@ class Table < Sequel::Model(:user_tables)
     self.new_table = true
 
     # finally, close off the data import
-    if data_import_id
-      @data_import = DataImport.find(id: data_import_id)
+    if @user_table.data_import_id
+      @data_import = DataImport.find(id: @user_table.data_import_id)
       @data_import.table_id   = id
       @data_import.table_name = name
       @data_import.save
@@ -595,17 +560,16 @@ class Table < Sequel::Model(:user_tables)
   end
 
   def before_save
-    self.updated_at = table_visualization.updated_at if table_visualization
+    @user_table.updated_at = table_visualization.updated_at if table_visualization
   end #before_save
 
   def after_save
-    super
     manage_tags
     update_name_changes
 
-    self.map.save
-    manager = CartoDB::TablePrivacyManager.new(self)
-    manager.set_from_table_privacy(privacy)
+    @user_table.map.save
+    manager = CartoDB::TablePrivacyManager.new(@user_table)
+    manager.set_from_table_privacy(@user_table.privacy)
     manager.propagate_to(table_visualization)
     if privacy_changed?
       manager.propagate_to_varnish
@@ -647,7 +611,7 @@ class Table < Sequel::Model(:user_tables)
 
   def create_default_map_and_layers
     m = ::Map.create(::Map::DEFAULT_OPTIONS.merge(table_id: self.id, user_id: self.user_id))
-    self.map_id = m.id
+    @user_table.map_id = m.id
     base_layer = ::Layer.new(Cartodb.config[:layer_opts]['base'])
     m.add_layer(base_layer)
 
@@ -669,9 +633,9 @@ class Table < Sequel::Model(:user_tables)
       name:         self.name,
       map_id:       self.map_id,
       type:         CartoDB::Visualization::Member::TYPE_CANONICAL,
-      description:  self.description,
-      tags:         (tags.split(',') if tags),
-      privacy:      PRIVACY_VALUES_TO_TEXTS[default_privacy_values],
+      description:  @user_table.description,
+      tags:         (@user_table.tags.split(',') if @user_table.tags),
+      privacy:      UserTable::PRIVACY_VALUES_TO_TEXTS[default_privacy_value],
       user_id:      self.owner.id,
       kind:         kind
     )
@@ -707,11 +671,9 @@ class Table < Sequel::Model(:user_tables)
     end
     @dependent_visualizations_cache     = dependent_visualizations.to_a
     @non_dependent_visualizations_cache = non_dependent_visualizations.to_a
-    super
   end
 
   def after_destroy
-    super
     # Delete visualization BEFORE deleting metadata, or named map won't be destroyed properly
     @table_visualization.delete(from_table_deletion=true) if @table_visualization
     Tag.filter(:user_id => user_id, :table_id => id).delete
@@ -830,55 +792,24 @@ class Table < Sequel::Model(:user_tables)
 
   def name=(value)
     value = value.downcase if value
-    return if value == self[:name] || value.blank?
+    return if value == @user_table[:name] || value.blank?
     new_name = get_valid_name(value, current_name: self.name)
 
     # Do not keep track of name changes until table has been saved
-    @name_changed_from = self.name if !new? && self.name.present?
+    @name_changed_from = @user_table.name if !new? && @user_table.name.present?
 
     self.invalidate_varnish_cache if !owner.nil? && owner.database_name
-    self[:name] = new_name
+    @user_table[:name] = new_name
   end
-
-  def tags=(value)
-    return unless value
-    self[:tags] = value.split(',').map{ |t| t.strip }.compact.delete_if{ |t| t.blank? }.uniq.join(',')
-  end
-
-  def private?
-    self.privacy == PRIVACY_PRIVATE
-  end #private?
-
-  def public?
-    self.privacy == PRIVACY_PUBLIC
-  end #public?
-
-  def public_with_link_only?
-    self.privacy == PRIVACY_LINK
-  end #public_with_link_only?
 
   def set_default_table_privacy
-    self.privacy ||= default_privacy_values
+    @user_table.privacy ||= default_privacy_value
     save
   end
 
-  # enforce standard format for this field
-  def privacy=(value)
-    case value
-      when PRIVACY_PUBLIC_TEXT.upcase, PRIVACY_PUBLIC, PRIVACY_PUBLIC.to_s
-        self[:privacy] = PRIVACY_PUBLIC
-      when PRIVACY_LINK_TEXT.upcase, PRIVACY_LINK, PRIVACY_LINK.to_s
-        self[:privacy] = PRIVACY_LINK
-      when PRIVACY_PRIVATE_TEXT.upcase, PRIVACY_PRIVATE, PRIVACY_PRIVATE.to_s
-        self[:privacy] = PRIVACY_PRIVATE
-      else
-        raise "Invalid privacy value '#{value}'"
-    end
-  end #privacy=
-
   def privacy_changed?
-    previous_changes.keys.include?(:privacy)
-  end #privacy_changed?
+    @user_table.previous_changes.keys.include?(:privacy)
+  end
 
   def key
     key ||= "rails:#{owner.database_name}:#{owner.database_schema}.#{name}"
@@ -1266,30 +1197,6 @@ class Table < Sequel::Model(:user_tables)
     save
   end
 
-  def self.find_all_by_user_id_and_tag(user_id, tag_name)
-    fetch("select user_tables.*,
-                    array_to_string(array(select tags.name from tags where tags.table_id = user_tables.id),',') as tags_names
-                        from user_tables, tags
-                        where user_tables.user_id = ?
-                          and user_tables.id = tags.table_id
-                          and tags.name = ?
-                        order by user_tables.id DESC", user_id, tag_name)
-  end
-
-  def self.find_by_identifier(user_id, identifier)
-    col = 'name'
-
-    table = fetch(%Q{
-      SELECT *
-      FROM user_tables
-      WHERE user_tables.user_id = ?
-      AND user_tables.#{col} = ?},
-      user_id, identifier
-    ).first
-    raise RecordNotFound if table.nil?
-    table
-  end
-
   def oid
     @oid ||= owner.in_database["SELECT '#{qualified_table_name}'::regclass::oid"].first[:oid]
   end
@@ -1408,21 +1315,17 @@ class Table < Sequel::Model(:user_tables)
     nil
   end
 
-  def privacy_text
-    PRIVACY_VALUES_TO_TEXTS[self.privacy].upcase
-  end #privacy_text
-
   # Simplify certain privacy values for the vizjson
   def privacy_text_for_vizjson
-    privacy == PRIVACY_LINK ? PRIVACY_PUBLIC_TEXT.upcase : privacy_text
-  end #privacy_text_for_vizjson
+    privacy == UserTable::PRIVACY_LINK ? 'PUBLIC' : @user_table.privacy_text
+  end
 
   def relator
     @relator ||= CartoDB::TableRelator.new(Rails::Sequel.connection, self)
   end #relator
 
   def set_table_id
-    self.table_id = self.get_table_id
+    @user_table.table_id = self.get_table_id
   end # set_table_id
 
   def get_table_id
@@ -1454,11 +1357,11 @@ class Table < Sequel::Model(:user_tables)
       propagate_namechange_to_table_vis unless errored
 
       unless errored
-        if layers.blank?
+        if @user_table.layers.blank?
           exception_to_raise = CartoDB::TableError.new("Attempt to rename table without layers #{qualified_table_name}")
           CartoDB::notify_exception(exception_to_raise, user: owner)
         else
-          layers.each do |layer|
+          @user_table.layers.each do |layer|
             layer.rename_table(@name_changed_from, name).save
           end
         end
@@ -1473,11 +1376,11 @@ class Table < Sequel::Model(:user_tables)
 
   # @see https://github.com/jeremyevans/sequel#qualifying-identifiers-columntable-names
   def sequel_qualified_table_name
-    "#{owner.database_schema}__#{self.name}".to_sym
+    "#{owner.database_schema}__#{@user_table.name}".to_sym
   end
 
   def qualified_table_name
-    "\"#{owner.database_schema}\".\"#{self.name}\""
+    "\"#{owner.database_schema}\".\"#{@user_table.name}\""
   end
 
   def database_schema
@@ -1531,12 +1434,8 @@ class Table < Sequel::Model(:user_tables)
 
   def update_cdb_tablemetadata
     owner.in_database(as: :superuser).run(%Q{
-      SELECT CDB_TableMetadataTouch('#{table_id}')
+      SELECT CDB_TableMetadataTouch('#{@user_table.table_id}')
     })
-  end
-
-  def update_updated_at
-    self.updated_at = Time.now
   end
 
   def get_valid_name(name, options={})
@@ -1570,7 +1469,7 @@ class Table < Sequel::Model(:user_tables)
 
     return name if name == options[:current_name]
 
-    name = "#{name}_t" if RESERVED_TABLE_NAMES.include?(name)
+    name = "#{name}_t" if UserTable::RESERVED_TABLE_NAMES.include?(name)
 
     database_schema = options[:database_schema].present? ? options[:database_schema] : 'public'
 
@@ -1647,7 +1546,7 @@ class Table < Sequel::Model(:user_tables)
     type = type.to_s.upcase
 
     self.the_geom_type = type.downcase
-    save_changes unless new?
+    @user_table.save_changes unless @user_table.new?
   end
 
   def create_table_in_database!
@@ -1704,10 +1603,10 @@ class Table < Sequel::Model(:user_tables)
   end
 
   def manage_tags
-    if self[:tags].blank?
+    if @user_table[:tags].blank?
       Tag.filter(:user_id => user_id, :table_id => id).delete
     else
-      tag_names = tags.split(',')
+      tag_names = @user_table.tags.split(',')
       table_tags = Tag.filter(:user_id => user_id, :table_id => id).all
       unless table_tags.empty?
         # Remove tags that are not in the new names list